enh api
