+++ index.html
@@ -28,7 +28,7 @@
   </noscript-->
   <div id="content" ng-controller="mainController">
     <div id="header-content" ng-include="'build/html/header.html'" ng-controller="headerController" onload="onLoadHeader()"></div>
-    <div id="stab-main-content" ng-include="'build/html/stab_main.html'" ng-controller="stabMidController" onload="onLoadStabMid()"></div>
+    <!--div id="stab-main-content" ng-include="'build/html/stab_main.html'" ng-controller="stabMidController" onload="onLoadStabMid()"></div-->
     <div id="main-content" ng-include="'build/html/main.html'" ng-controller="midController" onload="onLoadMid()"></div>
     <!--div id="about-content" ng-include="'build/html/about.html'" ng-controller="aboutController"></div-->
     <div id="footer-content" ng-include="'build/html/footer.html'" ng-controller="footerController"></div>
