+++ mock/book_post.json
@@ -2,6 +2,7 @@
   "parents": [
     {"id": "1","name":"南山大学"},
     {"id": "2","name":"名古屋大学"},
+    {"id": "2","name":"○○大学"},
     {"id": "3","name":"その他"}
   ],
   "categories": [
