+++ src/html/stab_main.html
@@ -1,7 +1,7 @@
 <!-- メインスタブコンテンツ -->
 <div>
   <h2>This content is Stab!!</h2>
-  <div class="waves-effect waves-light btn" ng-click="onLoadRequestAPI()" >Request API</div>
+  <div class="waves-effect waves-light btn" ng-click="onLoadRequestAPI()" >Request API from sample</div>
 
   <div>
     <div id="serch_count">検索数： {{serch_count}}<div>
